Fix `this.subject` for `moduleForModel` invocations without args.
