+++ __testfixtures__/ember-qunit-codemod/module-for-model.input.js
@@ -5,3 +5,11 @@ moduleForModel('foo', 'Unit | Model | foo');
 test('It transforms the subject', function(assert) {
   const model = this.subject();
 });
+
+moduleForModel('foo', 'Unit | Model | Foo', {
+  needs: ['serializer:foo']
+});
+
+test('uses store method', function (assert) {
+  let store = this.store();
+});
