+++ __testfixtures__/ember-qunit-codemod/module-for-model.output.js
@@ -1,11 +1,20 @@
 import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
 
+import { run } from '@ember/runloop';
+
 module('Unit | Model | foo', function(hooks) {
   setupTest(hooks);
 
-  test('It transforms the subject', function (assert) {
-    let subject = run(() => this.owner.lookup('service:store').createRecord('foo'));
+  test('It transforms the subject', function(assert) {
+    const model = run(() => this.owner.lookup('service:store').createRecord('foo'));
   });
+});
 
+module('Unit | Model | Foo', function(hooks) {
+  setupTest(hooks);
+
+  test('uses store method', function (assert) {
+    let store = this.owner.lookup('service:store');
+  });
 });
