--- __testfixtures__/ember-qunit-codemod/store.input.js
@@ -1,9 +1,0 @@
-import { moduleForModel, test } from 'ember-qunit';
-
-moduleForModel('foo', 'Unit | Model | Foo', {
-  needs: ['serializer:foo']
-});
-
-test('uses store method', function (assert) {
-  let store = this.store();
-});
