--- __testfixtures__/ember-qunit-codemod/store.output.js
@@ -1,10 +1,0 @@
-import { module, test } from 'qunit';
-import { setupTest } from 'ember-qunit';
-
-module('Unit | Model | Foo', function(hooks) {
-  setupTest(hooks);
-
-  test('uses store method', function (assert) {
-    let store = this.owner.lookup('service:store');
-  });
-});
